Fix #219
Don't set default for "slowDeviceMode" arg because doing so prevents "dotenv" from setting that value even if it's in your .env file (it's a key design criteria of dotenv to not override any existing value). The default value is already set in config.js, so we're fine just dropping it here.